remove asterisk from replaceIllegalFileNameCharactersInString
The regular expression used in replaceIllegalFileNameCharactersInString
is intended to remove illegal characters from filenames. However, the
regex pattern [\\,#%&{}/*<>$'":@\u2023|\\.]* is incorrect because it
includes an asterisk * which makes the preceding character class
optional, thus not effectively removing all instances of the specified
characters. This could lead to filenames still containing illegal
characters, which can cause errors when used in file systems that do not
support these characters. A correct approach would be to remove the
asterisk or adjust the regex to ensure all specified characters are
replaced.