fix: replace '?' in file names with empty string
It's only supported on Linux and MacOS, and from a quick look, there
isn't a nice way to differentiate between the desktop OS the user is on.
So just removing `?` seems to be fine. Another reason is that the `?`
create other issues on Unix systems due to `?` operating as a wildcard
operator.

If anyone is dissatisfied with this change, I'll look into alternative
solutions.